compute/reconcile: silence frontier reports for dropped collections
This commit adjusts reconciliation to prevent the replica from reporting
empty frontiers for collections dropped in the process. During
reconciliation, we drop collections/dataflows in two cases:

 1. When old dataflows are replaced by incompatible new dataflows.
 2. When old dataflows are not present in the initialization commands.

In both cases the controller does not expect to hear about the dropping
of the old dataflows, so the replica should be silent. This is now
ensured by changing the `reported_frontier`s of collections we are
going to drop to the empty frontiers. Doing so prevents the frontier
reporting logic to try and report the same frontier again when we allow
to compact away the old collections.

Note that case (1) mentioned above was already silent prior to this
commit, due to some special-casing in `report_dropped_collections`. This
special-casing has been replaced with an assert, as we no longer expect
`report_dropped_collections` to be called at all for collections dropped
during reconciliation.