compute-client: remove exceptions for expected spurious responses
This commit removes the exceptions we put in place, both in the compute
controller's response handling and the compute protocol spec, to account
for the fact that replicas could send responses for dataflows dropped
during reconciliation. Since the faulty behavior has been fixed now, the
exceptions are no longer required.