persist: rework JsonStats (#18984)
At interpret-time, we don't need detailed stats about cases where more
than one "category" of json datums (bool, numeric, string, list, map)
were present. In practice, these will always involve some sort of cast,
which would error on at least one of the categories. So, we rework our
json stats to be detailed if 0 or 1 categories are present, but replaced
with a sentinel if more than 1 are present. Note that this is not true
of JsonNull, so it's pulled out and always counted.

This is strictly less information than we used to keep, so we could add
a migration to convert the old format to the new, but this has only been
used on staging, so we save the complexity and pick a new proto id for
part stats.

Touches #12684