dbt-materialize: release v1.4.1
v0.52.0 inadvertently broke dbt-materialize v1.4.0, by disallowing
(most) `CREATE VIEW` statements on `mz_introspection`. Fortunately,
there is an easy workaround that we can release as dbt-materialize
v1.4.1, which is to rely on the new `auto_route_introspection_queries`
behavior in the dbt adapter, rather than explicitly setting the cluster
to `mz_introspection`.

There remains, I think, a glitch in dbt-materialize v1.4.1, which is
that it *requires* that you have a cluster named `default`, as the
cluster named in the connection profile is not actually used as the
cluster parameter for the session, but only as the default `IN CLUSTER`
cluster for any created indexes or materialized views.