storage: on boot, keep only the most recent source status history
Before this commit, the source status history was an append-only
shard, meaning it could potentially crash OOM any machine that
tried to read all of its contents.

Until we come up with a more principled solution aboout how
collections like this should work, truncating the collection on
boot seemed like a good idea. However, that seems potentially
annoying for support teams handling constantly rebooting clusters;
if we instead keep the most recent status from the last epoch, we
can maintain a bit of context as to what caused the crash.