environmentd: Add boostrap role to startup (#19008)
This commit adds the concept of a bootstrap role to startup. If a
bootstrap role is specified when starting Materialize, and it's the
first time that this instance of Materialize has been started, then a
bootstrap role will be created with the given name that has elevated
privileges. The role will have the `CREATEDB` and `CREATECLUSTER`
attribute, and the `CREATE` privilege on the `materialize` database, the
`materialize.public` schema, and the `default` cluster. The purpose of
this role is to make testing and local development easier when RBAC is
enabled.