doc: Refactor docs sidenav
This PR refactors the sidenav based on the information
architecture in #18508. For now, I've grouped existing
pages according to basic user goals, but we'll need
to write/rewrite various pages as a next step. For
example, under "Transform data", instead of linking to
the relevant SQL reference pages, we will want pages
explaining the concepts and how to best use them.

Addresses: #18508