docs/developer: focus design docs on recording assumptions, getting consensus, and recording decisions
We should do a light pivot where we focus on situations in which design
documents are used to:

 - record assumptions
 - record context
 - record decisions
 - write design documents for cross-team efforts

This means we should focus less on fully spec'ing an implementation in
the design document. Which in turn should lead to more succinct/targeted
design documents. The description of our process (modulo the removed
paragraph) is already aligned with these goals.

I also did shuffle around some of the "When to write a doc" bullet
points.