coverage: Don't print out unit-test-only and uncovered in one report
Because we have to add back lines for unit-test-only-covered, and remove
lines for uncovered code, the resulting lines might not be mapped
together line by line again as they were before. See
https://buildkite.com/materialize/coverage/builds/93#018804a2-19fc-4696-b11d-7e86383bd1f3
for an example of how this can go wrong, surrounding the line:

> debug!("sending pubsub message: {:?}", message);

Since the surrounding lines have a mix of being uncovered and
unit-test-only-covered the diff algorithm puts those blocks together
differently, and we end up with a hard-to-read coverage report. For now
I split this into two separate reports, one for uncovered code, the
other for unit-test-covered-only code. An alternative would be to
generate the output ourselves (inline in github comments or using
`buildkite-agent annotate`), `diff` is definitely not an appropriate
tool for this once you have more than one kind of information to be
shown.