orchestrator-k8s: enable Guaranteed QoS for pods
This commit makes the kubernetes orchestrator set both the 'limits' and
the 'requests' fields of all container resource configurations to the
same value. This is to ensure that, if limits are specified, kubernetes
assigns the pod to the `Guaranteed` QoS class [1].

A `Guaranteed` QoS class is a pre-condition for enabling the `static`
CPU management policy [2], which in turn is a pre-condition for enabling
CPU pinning for replica threads.

Note that we have to specify 'limits' and 'requests' also for the init
containers. If we don't, the pods resources are considered unlimited
[3].

[1]: https://kubernetes.io/docs/concepts/workloads/pods/pod-qos/#guaranteed
[2]: https://kubernetes.io/docs/tasks/administer-cluster/cpu-management-policies/#static-policy
[3]: https://kubernetes.io/docs/concepts/workloads/pods/init-containers/#resources