storage: multi worker postgres source
The original motivation for this work was to optimize the ingestion of
postgres sources by using all the available timely workers but in the
process many more improvements were realized.

Beyond the better performance characteristics this new implementation
improves upon the previous on the following additional dimensions:

- Error Handling
Previously each error could be tagged as
definite/indefinite/irrecoverable by attaching a suffix function call
like `.err_definite()` and requests to restart the dataflow had to be
manually wired up.

Transient error handling: This new work introduces a new operator
builder function (`AsyncOperatorBuilder::build_fallible`) which allows
using of the standard `?` operator to write fallible operators. This is
how all transient errors are handled and they all lead to the dataflow
restarting.

Definite error handling: A concrete `DefiniteError` enum type is
introduced with one variant per expected failure. There is no "catchall"
variant which enforces the necessary discipline around definite errors.
`DefiniteError` is the only kind of error that is allowed to end up in a
collection.

- Semantics of LSNs

In the code on main there were plenty of places where we were adding or
subtracting one to LSNs without justification. I have investigated all
the interactions with postgres to investigate how it interprets LSNs
(i.e if pg treats them as `<` or `<=`) and we know have the correct
conversions when the timely conventions differ from the pg conventions
with a comment explaining the behavior.

Additionally, I discovered that we were incorrectly using the
`commit_body.end_lsn()` to timestamp transactions instead of the
`commit_body.commit_lsn()`, which is always equal to
`begin_body.final_lsn()`.

- Separation of concerns

We previously had a very large async event loop with multiple labeled
`loop` statements where control flow was hard to follow. In this work I
have separated the handling of the replication stream in three distinct
components:

1. Extracting `LogicalReplicationMessages` while handling standby status
   updates (function `raw_stream`).
2. Extracting a single transaction from a stream containing
   `LogicalReplicationMessages` (function `extract_transaction`)
3. Alternating between reading the stream and discovering replication
   lag (top level loop)

Beyond that the snapshot and replication reading are separated into
their respective modules and their results are concatenated at the
timely level.

- Documentation

The code is now *extensively* documented with explanations and diagrams
of all the components. The main `postgres.rs` module contains a high
level overview of the dataflow and `snapshot.rs` and `replication.rs`
contain detailed descriptions of their mechanics.

- Succinctness

The code is now significantly shorter and (in my opinion) simpler to
follow. The diff count on Github doesn't really do it justice due to the
hundreds of lines of documentation added.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>