[persist/expr] Fix error-swallowing behaviour in the expression interpreter (#19173)
Previously, the interpreter
[assumed](https://github.com/MaterializeInc/materialize/pull/19173/files#diff-a5a7da55d74ef805375fd6b3fe734fd2fd262f01f9d2642aee8d7c09d5c5ed8dL368)
that a function call would fail if evaluating any of its arguments
fails. This is mostly true, but is wrong at least in the case of `and`
and `or`, which can return a result even when some arguments evaluate to
an error.

Instead of special-casing those functions, this generalizes the
`flat_map` function on `ResultSpec` to handle `Result`s and not just
`Datum`s. The main downside of this approach is that we can't pass the
arguments as columns any longer, and passing them as literals in the
expression involves heap allocation. The last commit mitigates this
slightly by reusing the allocated expression between calls to `eval`.