persist: extract a first-class columnar struct abstraction
And migrate our `Schema<Row>` and `Schema<SourceData>` implementations
to use it instead of RelationDescHack. Ideally this would have
been two commits/PRs, but removing `Data::TYPE` makes it tricky.

Most of the actual logic already existed in `mod part` (duplicated once
for key and once for val, in fact) so, at a very high level, this is
extracting that into a nice abstraction.

We can't infer the schema of a `Row` from the type (as opposed to some
hypothetical `Row<usize, String, Vec<i32>`), which is the first time
that's true for a `Data` impl. This means we need a bit of a
generalization of the model. Concretely, the 1:1 relationship between
variants of `enum DataType` and impls of `trait Data`. However, we add a
new `type Cfg: ColumnCfg` associated type to `Data`, which may affect
the columnar schema at runtime. For all previously existing `Data`
impls, this is unnecessary and so we use `()`. For the new `DynStruct`
impl of `Data`, we use a new `DynStructCfg`.

This new Cfg replaces the old `const TYPE` on `Data`, which affects a
bunch of the plumbing (e.g. `Data::Mut` replaces the `Default` bound
with an impl of `fn new(cfg: T::Cfg)` on a new `ColumnMut` trait, which
mirrors `ColumnGet`). This is the bulk of the diff.

Notes:
- We use reuses `DynStructCol`/`DynStructMut` as the `Data` impl for
  both `DynStruct` and `Option<DynStruct>`. It would be more type safe
  to use different ones, but I couldn't find a great way to do this
  without the boilerplate exploding even more than it already has here.
  There's already precedent for this in the `Data` impls for `String`
  and `Vec<u8>`.
- This introduces the ability for a column of structs to be used via the
  normal `Data` interfaces, but somewhat sadly, `Row` and `SourceData`
  still reach inside and directly manipulate the columns. Doing
  otherwise would be a pretty significant perf hit. It might be possible
  to fix this up by making DynColumnRef and DynColumnMut into enums
  (essentially moving the DatumEncoder/DatumDecoder enum_dispatch
  pattern into the columnar abstraction). I played with it for a bit and
  it's a considerably bigger change than even this, there's one or two
  tricky bits still to work through, and I've spent far too much time on
  this PR already :).

Benchmark for Row confirms performance hasn't regressed:

    roundtrip_encode_structured
                            time:   [739.79 µs 740.11 µs 740.49 µs]
                            change: [+0.2556% +0.4244% +0.6203%] (p = 0.00 < 0.05)
                            Change within noise threshold.

    roundtrip_decode_structured
                            time:   [627.30 µs 627.58 µs 627.93 µs]
                            change: [-1.7527% -1.5683% -1.4115%] (p = 0.00 < 0.05)
                            Performance has improved.