mzcompose: Principled handling of lists of prior versions
Previously, it was not possible to distinguish between the
version list as maintained in the docs and the version list
maintained as Git tags.

Furthermore, it was not possible to distinguish between minor
and patch versions, which is important when targeting specific
upgrade sequences.

This commit refactors mzcompose.util.get_released_versions() into
a set of classes and functions that allow those distinctions to be made.

This commit does not change the behavior of the existing upgrade tests
but a follow-up commit will take advantage of this refactor in full.