doc: Postgres direct source integration guides
This PR provides end-to-end guidance for ingesting data directly from
PostgreSQL, including db setup, network security options, and post-connection
verification. The network security options are reflected in the Materialize
object creation steps, as appopriate.

Whereas previously we had a single integration guide with tabs for different
PostgreSQL services, we now have a dedicated page per service. A lot of the
content is similar across guides, but we use Hugo shortcodes to reduce
some redundancy, and this new approach is friendly to users at the expense
of some added maintenance complexity for us.

This PR also changes "Connnect source" to "Ingest data" in the sidenav,
updates the PostgreSQL topics on the integrations overview page (removing
mention of DigitalOcean Managed PostgreSQL, CrunchyBridge Postgres,
Heroku, and Aurora Serverless, with Devex approval), and updates links
throughout.

Fixes: #19038, #18764, #14871