[persist] Handle MFP's strict expression evaluation. (#19378)
In #19173, we taught the `ColumnSpecs` interpreter to support AND and
OR's non-strict evaluation behaviour.

As part of evaluating an MFP's filters to determine whether it will
filter out an entire set of rows, we use `AND` to combine the results of
multiple filters. However, MFP's filter evaluation is strict: if any
filter errors the overall `evaluate` call will fail. This caused the
results of evaluating an MFP with `eval` or the new interpreter to be
inconsistent.

Instead of re-implementing AND in a strict way, this changes the
abstract interpreter to permit both semantics: now an expression like
`<error> AND false` is allowed to _either_ error or return false. Aside
from fixing the linked bug, this also makes the interpreter more robust
to future changes in the optimizer.