Improved Exchange hashing in `consolidate_named_if` (#19372)
This commit fixes an observed issue of poor balance among workers in
conditional consolidation of some monotonic one-shot SELECT queries.
We observe that the quality of hashing in Exchange influences
our ability to achieve good parallelism in consolidation by ensuring
proper distribution of rows to workers (#18890).

To address the above problem, this commit proposes the use of AHash
in the Exchange of consolidate_named_if, since AHash has empirically
good quality and speed in addition to being adopted already in our
dependencies (namely, arrow2, hashbrown, and kube-runtime). Also, using
AHash, we observe good balance among workers in queries where this
was an issue with DD's default FnvHasher, while little to no degradation is
observed in other queries (#19372).