storage-client: introduce alter_compatible
Given the complexity of the ingestion-related structs, we should
force transformations to be explicitly allowed, rather than
blithely accepted.

By placing this in the storage-client, this also enforces a clear
abstraction boundary with the adapter for upcoming ALTER SOURCE
work.

Note that this commit allows keys to be added and removed from
map fields, even though there is currently no way of actually
performing those changes.