persist: pass down struct validity when computing column stats
In arrow, nullable structs are "dense"; each component column of the
struct must have a value at each offset, even if the struct is null at
that offset. This means that, for a nullable struct with a non-nullable
field, we end up with placeholders in the field column. This must be
ignored when computing statistics else we'll end up with incorrectness
(unnecessarily wide stats) or panics.

The arrow2 library addresses this by given every Array a validity,
whether it's nullable or not, and then nullability is "emergent" at
runtime.

We solve the problem the opposite way, by passing the nullable struct's
validity down when computing stats (recursively for a struct of
structs). To make it slightly easier to work with, we declare a new
invariant (that we were already keeping) that for a nullable field in a
nullable struct, the validity of the column must be a subset of the
validity of the struct.

This is a decently big footgun, but it's not clear that anything else is
obviously better.