github: Update pull request template
Previously, one of the checklist items in the pull request template
would only be checked conditionally for certain types of pull request.
All other checklist items would be checked for all pull requests. This
left most pull requests with all but a single check list item checked.

This commit updates the wording of the conditionally checked checklist
item so that it can always be checked. This will help make all pull
requests consistent in the amount of checklist items they should check.