Revert "cargo-test: don't run with more CPU cores"
This reverts commit 7449de7f7412098cff15fce4365aa7f5fb391ded.

Running with more CPU cores wasn't the problem. The problem was that
the `clusterd` binary was not being copied to the right place.

Just revert that part of 8cd0b1b rather than fixing the path. I'm almost
certain that building `clusterd` is substantially faster than trying to
copy it out of the `Clusterd` image. We have to build the test binaries
in this job no matter what, so building `clusterd` too is relatively
immaterial, and not having to wait for the `build_x86-64` job makes
cargo-test results available much sooner for the first build.