Design doc to stabilize monotonic one-shot `SELECT`s (#19250)
This commit adds a design document focusing on the steps necessary for
stabilizing monotonic one-shot `SELECT`s. The design document covers
impacts on `EXPLAIN` and CI testing, as well as further optimizations to
refine consolidation when monotonic operators are introduced in
single-time contexts.

---------

Co-authored-by: Jan Teske <jteske@posteo.net>
who provided typo fixes to some of the formulations in the text.