Migrate the Stash to Protobuf (#19373)
### Motivation

Fixes #17466

This is the final PR for improving Stash migrations. The general
approach we're taking is to migrate the Stash from having serialized
JSON to serialized protobufs. We define the objects we store in a
`.proto` file, every time we want to change this file we must first
snapshot it (i.e. copy and paste) and write a migration from the
previous version to the new version.

All fields in proto3 are optional by default, without any way to mark
them as required. To make working with Stash collections easier, we
re-use the existing types in the Catalog, implementing
`mz_proto::RustType` for them to do the conversion between proto and
Rust. I also updated `TableTransaction`, which is widely used in the
catalog, to best encapsulate this conversion. While this introduces a
bit of overhead, it decouples our application layer from our storage
layer, which gives us some additional flexibility with how data is
stored in the Stash.

### Tips for reviewer

It's easiest to review this PR one commit at a time. I did my best to
break it down into logical sections, the first being the largest.

1. Implement `mz_proto::RustType` for a bunch of types. 
1. This is straight converting from one type to another. Previously I
had added a `StashType` trait but realized the already existing
`RustType` had exactly the same intent, and was better in many ways. So
this commit involves deleting `StashType` and implementing `RustType`
for those types, as well as a number of new types.
2. Changes the Stash itself and the inner postgres workings to use bytes
and protobuf.
3. Migrates the Catalog to use protobuf types.
4. Migrates the storage-controller to use protobuf types.
5. Ports the existing Catalog/Stash migrations to the new framework
6. Updates the `stash-debug` crate to use protobuf types.
1. Note: making edits with stash-debug are now quite difficult since we
can’t take a “generic protobuf” as we could take a “generic json blob”.
I was thinking to improve this we could use
[`prost-reflect`](https://crates.io/crates/prost-reflect) so we could
take an arbitrary string as the name of the collection, and then
dynamically prompt the user for the fields in the message. Making this
change is out of scope for this PR though.
7. Updates the Stash tests, and makes a few correctness changes to the
initialization flow.

### Testing

#19208 introduced the actual migration flow. In that PR there are a
number of tests to exercise scale and different errors paths. For this
PR I'm more so concerned about how the protos work during normal usage
of the Stash. As such, I was planning on running the
[nightlies](https://buildkite.com/materialize/nightlies) and [SQL Logic
Tests](https://buildkite.com/materialize/sql-logic-tests) on this
branch.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
[template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md),
or is sufficiently small to not require a design.
<!-- Reference the design in the description. -->
- [ ] This PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way) and therefore is tagged with
a `T-proto` label.
- [ ] If this PR will require changes to cloud orchestration, there is a
companion cloud PR to account for those changes that is tagged with the
release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
    - N/a